Moved the composite schemas to the top level of the project
RMB 19 has documented difficulty with relative paths that use '..'. In this case, we were using '../..' which, on UNIX, leads to files referenced this way not being found by RMB. The solution is to move these types of files to the root of the project. This is similar to compositeUser in the raml project in the schema directory instead of schema/mod-users-bl, where it is used. If it was in mod-users-bl it would have required '..' in its refs. The docs claim 1 level of '..' is fine, but not more than one, which is the case we were in.